pyproject.toml
[tool.poetry]
name = "hubquest-taxi"
version = "0.1.0"
description = ""
authors = ["Dawid Kowalczyk <dawid.kowalczyk@iff.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
haversine = "^2.8.0"
scikit-learn = "^1.3.2"
numpy = {version = "^1.25.2", python="^3.10,<3.13"}
pandas = "^2.1.2"
pyarrow = "^13.0.0"
holidays = "^0.35"
pyaml = "^23.9.7"
plotly = "^5.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
